Better socket termination for ZeroMQ
Sending and closing ZeroMQ sockets only
block on access to the file descriptor,
not on the sending of data or the closing
of sockets.

As such, the timeout in _cast() was ineffective and
resulted in immediately calling close().

Because close() defaults to linger=-1, the underlying
file descriptor (managed by a ZeroMQ thread spawned
within the C library) would only be closed
upon the successful sending of the message.

This would result in a file descriptor leak and
infinite attempts to deliver messages despite
the timeout per message sent to a non-existant consumer.

The solution is to use the linger
argument to expire the messages if they cannot
be delivered within the timeout window.

Adds a tests to test_zmq that assures messages cast
to a topic cannot be consumed after the expiration
of the cast timeout and, to complete coverage, of
the new code, that messages may generally be
consumed following a message cast.

Fixes bug 1193439

Change-Id: I07c7996ee76873ee762582e58dbbec05bc7de66c