Return group name 'unknown' for options with no group info
Rather than raising an exception, print a warning that
the group name of an option was not found. For options
that were registered twice or not registered on the same module
it was defined should be fixed manually. There are some options
which its groups are created at runtime therefore impossible to determine
the groupname. For instance, the cinder lvm options are set on different
backend names, i.e

[lvm1]
volume_group=lvm-group-1
...

[lvm2]
volume_group=lvm-group-2
...

Change-Id: I3bcb506a2366d616de8f390997a64725307d332d