Add decorator to make ABCs out of classes.
Rationale:

Creating abstract base classes in a way that works correctly for both
Python 2 and Python 3 has not yet been addressed. This module adds a
simple decorator that can be applied to any class and sets the
metaclass of that class to be abc.ABCMeta.

Tests are provided that demonstrate expected behavior for all
supported platforms.

*Note*: Running the tests on Python 3 took some finagling. I had to:

1. Create a py3.3 virtualenv
2. Removes several deps from requirements.txt and test-requirements.txt
3. Install nose.
4. Run `nose tests/unit/test_abcutils`

Until python 3 is working for all of the oslo dependencies, running
automated tests on abcutils is going to be a challenge.

Change-Id: Ib7074cf3777529afb8f100da1dc87516b4cd8011