stop using + \ in help string
Hacking specifies 'Long lines should be wrapped in parentheses
in preference to using a backslash for line continuation.' Was
noticed when merging this into nova. It's a minor cleanup, but
good to be consistent.

This keeps the old formatting of no newlines, though if newlines
were fine a docstring might also be appropriate.

Change-Id: I8adccdc8bf82f7b97dbbc8969d6b9de4be31e214