Adds common image utils
This is the first step to merge the code in nova/virt/images.py
and cinder/image/image_utils.py.

This patch pulls up the QemuImgInfo class and the qemu_img_info
function which are common in the two modules above.
The test script was taken from nova/tests/virt/test_images.py with
an additional tearDown function and test_qemu_info_file_absent
testcase.

Change-Id: Ia07739492776180e579b03920aa32b9215ce941d