Revert "Implement apiclient library"
This reverts commit 062cc24f5dfc8e1652a589f6a2f45ce8f3a1b89c.

The apiclient is a very important idea.  However, that authentication
portion is highly likely to have security updates required in the
future, and having a copied section of code in multiple client
libraries will make patching difficult.

The implementation for this approach has been targetted for the
keystoneclient library, with the understanding that the keystoneclient
will then be used by the other clients to perform authentication.

The code as commited here has not yet been vetted in Keystoneclient.
There are some identified changes that will have to happen before it
is integrated there.  Thus, the version of code that is included in
this repository would have been incorrect very quickly.

The one change not removed with this revert is the addition of the
requests library to requirements.txt.  Removing that change now will
cause test failures elsewhere.

Change-Id: I593316dd3a722e6fdcb38409be4a2f0ca3bb3caf