requeue instead of reject
Requeue the message instead of rejecting it. The AMQP spec
suggests reject()'ing the message to get it back in the queue.
The trick is that there is a parameter "requeue=True" required
to make this happen. Under kombu, this is done as a separate
command instead of a parameter on reject().

Change-Id: I6e820841b87eedb03318124e7d70be8da6970cc2