Add test for checking models with migrations
Each component of OpenStack uses models as base for service
and migrations as base for creation of database.
In fact there are a lot of differences in models with migrations.
And there is no test that could find them.

This patch adds opportunity to compare actual state of DB with models
in mysql and postgres dialects.

Will be checked:
- tables with shadow_tables,
- columns (name, type, lenght, default, nullable),
- constraints (indexes, unique constraints and foreign keys).

Also there is a alembic support for project which use alembic repo
as base or auxiliary.

bp: sync-models-with-migrations

Change-Id: I415e11864a1859cf074e4e55eb610ae8ce96e8c5