Proper usage of _resource property in quotas
Use resources instead of _resources. This is necessary because _resources
actually does not contain all resources in VolumeTypeQuotaEngine, and there
is property-method `resources' was introduced for this.

Use `self.message' attribute in QuotaError constructor. This attribute can
be used for access to exception's message in except block as done in cinder.

Change-Id: Iac85497a956e7c3a8c58bb38337a63d0d5fabc8f