Update requirements
This is the fabulous result of running the update.py script from
openstack/requirements with this repository as the target. It therefore
updates all dependencies, with nice version numbers that this repository
have to use to work. It also fixes case on some dependencies.
This update also brings a brand shiny new setup.py automatically handled
and generate by this same openstack/requirements repository.

Due to some changes somewhere, testr configuration has to be modified,
and I actually fixed and simplified a bit to make tests ran.

Change-Id: Ibbea49fbc98caa6a175e9174243ab66a753493f8