Clean up db.sqla.Models.extra_keys interface
Previously we were checking that Class has
_extra_keys() method and if it has then extend columns with this list.

It will be much cleaner to have in ModelBase this method so we will be able
to avoid this check,  and as method should contains verb we change the name
also to _get_extra_keys().

It changes API so we should also make related in other projects.
But there is not so much work (e.g. in Nova there are only 2 occurences)

Change-Id: Ib9f3a165fae340a1c887769ce8f21a70c28458d0
Closes-Bug: #1119717