Change Message to extend built-in text type
The original fix had misdiagnosed the underlying problem. For
background, this is how strings are processed by logging in py27:

1. The message object to be logged is added to the log record via
the "msg" attribute.

2. logging.LogRecord.getMessage does an isinstance(msg, basestring)
check.  If negative, the value of the "msg" attribute is replaced
by str(msg).  This was happening with openstack.common.gettexutils.
Message objects, since they did not extend either unicode or str
(py27).

3. The logging formatter combines metadata and the message to be
logged with the formatting string via a mod operation. If any of
the mod parameters are unicode objects, all of the other parameters
are promoted to unicode objects as part of the mod operation. If
the Message object in step 2 contains double-byte characters and
the default encoding is 'ascii' (which it usually is), the
promotion of the str object that was derived from the Message
object fails because it cannot process any non-ascii bytes.

Because the logging module and other string-processing modules
cannot be changed to accommodate Message objects, the only
alternative was to alter the Message class so that it extends
the built-in text type (unicode in py27) for double-byte
characters.

The implications of this change are that messages are no longer
translated when they are rendered; they are translated when
they are created and, from the user's perspective, they are
immutable, because the built-in text types are immutable.
The way to translate a Message instance is to call
its translate_into method; the Message instance that is returned
is the translated message, which itself can be translated.
In other words, translating a Message object does not alter
the underlying message ID.

closes bug 1225099
fixes bug 1229800

Change-Id: I9b67e57f4f9bea2883e76c7b20651879770d47bd