Change Message to extend built-in text type
The original fix had misdiagnosed the underlying problem. For
background, this is how strings are processed by logging in py27:

1. The message object to be logged is added to the log record via
the "msg" attribute.

2. logging.LogRecord.getMessage does an isinstance(msg, basestring)
check.  If negative, the value of the "msg" attribute is replaced
by str(msg).  This was happening with openstack.common.gettexutils.
Message objects, since they did not extend either unicode or str
(py27).

3. The logging formatter combines metadata and the message to be
logged with the formatting string via a mod operation. If any of
the mod parameters are unicode objects, all of the other parameters
are promoted to unicode objects as part of the mod operation. If
the Message object in step 2 contains double-byte characters and
the default encoding is 'ascii' (which it usually is), the
promotion of the str object that was derived from the Message
object fails because it cannot process any non-ascii bytes.

Because the logging module and other string-processing modules
cannot be changed to accommodate Message objects, the only
alternative was to alter the Message class so that it extends
the built-in text type (unicode in py27) for double-byte
characters. This change entails some complications...

These built-in text types are immutable, while the Message
objects were coded to be mutable in terms of the locale. The
relevant methods and test cases had to be altered to accommodate
the new, partially immutable objects.  This was necessary,
because some operations will obtain the string data from the
immutable unicode data, while other operations call the
__unicode__ method instead.  Every change to a Message object
that alters the value returned by __unicode__ method therefore
requires that a new Message instance, one with updated
unicode data, be created.  From a consumer's perspective, the
only affected functionality is translating the string into
a different locale.  This is now done via the Message.translate_into
method, which will return a new Message object with the corresponding
updates to the immutable unicode data. (Consumers should no longer
alter the locale property directly.) Note, however, that the
underlying message key and other instance data remains unchanged,
so that a message that has been translated into a different locale
can be translated again--no information that is necessary to
render a new translation is lost.

closes bug 1225099
fixes bug 1229800

Change-Id: I9b67e57f4f9bea2883e76c7b20651879770d47bd