Enables db2 server disconnects to be handled pessimistically
Fixes bug 1231657

A checkout listener can be added to sqlalchemy to be able to
handle database disconnect pessimistically and avoid an error in
the event of a DB restart or simply a temp connection issue due
to network glitches. At the moment only mysql is enabled with
this feature. This patch enables db2 as well.

Depends on change I7f91ab1978399f6dcabc76c939be54746db14620

Change-Id: I5563360b43c4b648f5b035099b0e1d2e71d32503