Add LockFixture to lockutils
This commit adds a fixture class to be used instead of external
locking with the synchronized decorator. When running tests in
parallel the synchronized decorator won't synchronize the tearDown
and addCleanup methods for a test. By using the fixture which does
LockFixture the unlocking is done by using addCleanup() ensuring that
tearDown and later addCleanup methods are run inside the lock
context.

Closes-Bug: #1229437

Change-Id: I61af5e9d385bb7ef2e6402366fb8c5c09f749986