Revert "Make Messages unicode before hitting logging"
This reverts commit 572cfecb4da5c618a407b8de2a7feae9c3f35ce4.

The commit being reverted was sent as a workaround for bug #1225099
however the translation features were ultimately completely turned off,
and so the workaround ended up not being needed and it is actually
unnecessary code.

We will try to find a better solution for the bug and we'll send it when
the translation features are turned back on.

Change-Id: If690d8ba6351bcacd77fe15454b2c536e3213273