Don't override default value for eventlet.wsgi.server logging
We want eventlet.wsgi.server to use the same log level as everything
else (by default). For example if INFO logging is enabled eventlet.wsgi.server will
log all API requests. So if INFO logging is enabled in the API server we
want eventlet.wsgi.server to use the same log level.

This only changes a default value, and can be set by the user.

Change-Id: I42124793bbf542d69f6493bb3a9611e443775c2a