Fix cleanup ordering in moxstubout fixture
The ordering of fixture cleanup calls is very important in preventing test
interactions.

Originally, Nova got this ordering wrong, so we had to fix it here:

    https://github.com/openstack/nova/commit/4e560d49d4a8d08341dae6df5e123e1be1b3c779

Unfortunately, when the fixture was added to Oslo, it was added using the
original incorrect ordering, not the revised, correct ordering.

This caused Nova to regress and suffer the same sporadic test failures we had
when fixing the original Nova bug.

A recent commit

    https://github.com/openstack/oslo-incubator/commit/f4a485530ebd123cdb80b8b8a378ff9847b2a5d7

attempted to fix this by sharing stubs between the self.stubs and self.mox.

Unfortunately, this caused other Nova tests to fail
(nova/tests/network/test_manager.py).

The fix is to use the original Nova fixture with the correct cleanup ordering.

Fixes bug 1239898

Change-Id: I0971eacf165667bb060b5cd7ec4f8182d71346e1