Allow keeping of existing loggers with fileConfig
Currently, common.log.py setup() will disable all existing
loggers that were set up before importing a custom logging
configuration from file. This results in a lot of modules
not having any logging feeding into the root logger that
is overriden to redirect log output in the custom setup.

This change adds a config flag to allow users to change this
behavior and keep the existing loggers so they will be properly
redirected to some output mechanism specified in the user config.

Closes-Bug: #1169328

Change-Id: I1f25300d44b04ca5bcdd9b505319f0d089a9c964