Revert "Allow non-use of cfg.CONF in securemessage"
This reverts commit 5c71c257e001b16e03ab78be4518736d5256c3bd.

I don't have enough details about why we did this but it is
breaking some things... namely the config file generator
does not work properly when we lazy load/init our config
file groups like this. We could fix the config file
parser so it ignores things in this case... but it seems reasonable
to revert this too...

I hit issues w/ this because I'm trying to add some missing
crypto modules to Nova in:

 https://review.openstack.org/#/c/52643/

This causes the config file generator to blow up with the following
message:

Unable to find group for option enabled, maybe it's defined twice in the
same group?

Closes-bug #1241720

Change-Id: Ia09bb82cb65aa280d64c5bccaca62e983a6f7a2d