Guess the group name for Lazily registered options
securemessage.py for example lazily registers options into the
'secure_messages' group. Current code in generator assumes that all
registration is done on the global CONF. In this review, we relax
that a bit and looks if there is only and only one OptGroup is
in the module as well and assumes that there would be a lazy
registration of the options into that OptGroup sometime later.

Change-Id: I8aaa1267e32fd1080318031945fd0f7594bd6250
Closes-Bug: 1241720