Drop RPC securemessage.py and crypto module.
The RPC securemessage stuff will eventually get integrated into
oslo.message. As is it is only partially complete in oslo-incubator
and it is getting synced into various projects (and breaking
some things).

This commit drops securemessage.py and the associated crypto module
since securemessage.py was the only user.

Related-bug #1241634

Change-Id: Ie846dc01c20e44cc6cc4b4e910665f486749d317