Fix to support ipv6 in sql connection urls
sqlalchemy has trouble with ipv6 urls, so we should parse
the url ourselves before we call sqlalchemy.create_engine.
The regexp can get us part of the way and we can use the
parse_host_port method to extract the host and port.
Added a bunch of tests with sample urls found in the
logs and our code base.

Change-Id: I4bb3a23dee70e33b2bd931eb203b7d12617f2849
Closes-Bug: 1190454