Allow generate_sample MODULEPATH to be set in env
This commit allows the default modulepath to optionally be set with
an environmental variable. This is useful for tempest because config
isn't used in the same way because tempest is executed in a test
runner. So to enable using generator.py it has to be wrapped to
manually load the config before calling generate(). This will enable
specifying the local tempest generator wrapper module without modifying
the oslo synced script.

Change-Id: Iacef39d6ae7693ecd09b113401b7db52afc53d17