Added opportunistic DB test cases
These test cases are meant to be run opportunistically within the
general test suite, i.e. only when openstack_citest database is
available (which is always true for our CI infrastructure, but might not
be true for developers machines). This is similar to how we run
migrations tests in all OpenStack projects

Blueprint: tests-given-db-backend
Change-Id: Ic3706763c8fc56c0fcf4d1964ed1c5f17a662be6