Translation Message improvements
The improvements to the current Message implementation include:

* Implementing a new method translate() to explicitely perform the
translation of a Message object, instead of relying on setting the
locale property.
* Remove inheriting from UserString and stop trying to make the Message
class act like a string, because that was causing problems with
consumers (e.g. the python log formatter) pretending the Message was a
string when it really was not.
* Remove the internal representation of the translated Message (i.e. the
data property) and simply return a unicode translation upon translate()
* Banning of the str() and unicode() methods, requiring users to always
use translate() to use this class.
* Enhancement to the corresponding TestTranslation class to enable fake
translations for multiple locales.


Partially implements: bp i18n-messages
Change-Id: I2dbddf422e79fd4474dede1e7c7ed229c89dc30c