Translation Message improvements
The improvements to the current Message implementation include:

* Implementing a new method translate() to explicitely perform the
translation of a Message object, instead of relying on setting the
'locale' property.
* All of the codebase assumes _() is a unicode object, so make the
Message unicode, that will always behave as unicode, except for the fact
that it has the extra ability to translate() on demand.
* Remove the internal representation of the translated Message (i.e. the
data property) and rely on parent implementations.
* Enhancement to the corresponding TestTranslation class to enable fake
translations for multiple locales.

With this patch the REST API translation in the other projects will work
with very minimal or no impact after re-enabling the 'lazy' _(). 

Partially implements: bp i18n-messages
Change-Id: I2dbddf422e79fd4474dede1e7c7ed229c89dc30c