Improve error reporting on import failures
The existing config generator code does not report general errors
importing modules well.  In my case, that meant I had no idea that
an old version of six on my system was causing a module import to
fail.

This change adds better error reporting and cleans up a problem
that was causing an unnecessary import exception.

Change-Id: I4663f5bbd4f553c0d6ce5f84b1c1275a7304e3b5