Preparing to use different backends
Assuming that  various backends will be used a cleanup is necessary.
Real sql databases are not self-destructable thus it is need to be cleaned.
Moreover we must be shure that an `engine` instance is newly created one
and connected to a correct backend. So session must be cleaned too.
Otherwise `get_engine` will return cached instance with potentially
unexpected connection.

Change-Id: I81d1ac2804dc7434f7ed1f36206f39f777480398