Have tox install via setup.py develop
tox 1.6 was released, which means that we can now take advantage of the
feature we added to it - which is using setup.py develop to install the
code into the virtualenv. The logic was taken from run_tests.sh - so the
performance issues around using tox vs. using install_venv should now be
gone.

Ported from Nova:
https://github.com/openstack/nova/commit/32d63a9fbe6f353ca1af8922d6bcbb012a2cc32c

Change-Id: If83a5b5473344d3a0e3a3a9fbb4ba16fa75f5be3