WIP: attempt to see if changing queue names makes parallel work
this is a crazy experiment to see if just changing the queue names
to be non overlapping is all that's needed for parallel. Mostly
because getting zmq tests to work locally was a fail, so see what
the check queue thinks. If this looks promissing a more complete
patch (and better commit message) will come later.

Change-Id: I46eec5977729ac2565d7e20a8c35ded9df5fdf36