provide a global default for lock dir
this is based on one proposed upgrade path for cinder on the
openstack-dev mailing list. Instead of making oslo lockutils error
if the user has not added a config variable for lock path in their
service (which they never needed to do before) use tempdir instead.

in order to address the security concerns of file truncation attacks,
open the files for locking in 'a' mode, which will make them non
destructive to the underlying files. Add tests to ensure that we
don't truncate underlying files.

also throw an error message to the user that this isn't a safe
mode to run in. That will hopefully mean they will only use this
for the small window of rolling their cluster forward, then
configure it to something more sane.

(note: this patch could use tests for that log message, comments
about how to mock that correctly in oslo would be appreciated.)

Change-Id: Ia96a8926741464cf0d47d2c25c77b9c746c1b334