Change IgnoreAttemptedHostFilter to expect 'retry' key
The newly added IgnoreAttemptedHostFilter is renamed from RetryFilter, but
there's also a change in the key it expected in filter_properties ('retry'->
'attempted').  This change actually breaks backwards compatibility if it's
been used in an environment where old/new scheduler worked together and old
scheduler processed the request prior new scheduler (old scheduler that uses
RetryFilter will add 'retry' key to filter_properties).  This kind of
situation can happen in a rolling upgrade in a multi-scheduler deployment and
should be fixed.

Change-Id: I0da5092833190f7e6dadfaaacfa4545e85d45a15