Fixing misspelled encryption key in QemuImgInfo
This modification fixes Bug #1257829, specifically addressing the use
of a misspelled 'encryption' key to retrieve image encryption
information returned by calls to the 'qemu-img' command. In addition
to fixing the key ('encryption' -> 'encrypted'), the QemuImgInfo class
attribute has been renamed to reflect the change and the encryption
information added to the string representation of the class. Test
cases verifying the parsing of the encryption information are also
included.

Change-Id: Ice087d95c52506dec735921c6e9477caf082bf7e
Closes-Bug: #1257829