Allow configurable ProcessLauncher liveness check
The current interval between ProcessLauncher checks of whether
a child process has exited - 0.01s - prioritizes fast respawning of
children at the expense of cpu usage.  This tradeoff may not be
appropriate for all clients of the class, so this patch makes the
interval configurable.  The current interval remains the default
to avoid changing expected behavior.

This patch was motivated by unnecessary idle cpu usage of the
Neutron service when configured to use ProcessLauncher-managed
worker processes.

Partial-Bug: #1095346

Change-Id: I2f5e06f98be1f75c21e6681967b573f5d88a4ed6