disable SIGHUP restart behavior
SIGHUP should only be a restart action for a daemonized processs,
when used on a console process SIGHUP should == SIGTERM. However
there is no detection about whether this service is actually
daemonized.

This change makes it so that SIGHUP == SIGTERM until someone
implements the detection. Until this is merged (and synced with
nova) unstack.sh is broken, and not really fixable on the devstack
side. This will impact future upgrade testing.

Change-Id: I2a73d3872a302fcfac1a6ee844a44faa9569b3dd
Partial-Bug: #1263122